+++ src/small_tree.h
@@ -141,7 +141,7 @@ struct small_tree_group : public vector<small_tree> {
 	void maybe_add_tree(int i, int j, float zpos_in, float tsize, int skip_val, bool check_hmap_normal);
 	void gen_trees(int x1, int y1, int x2, int y2, float const density[4]);
 	void gen_trees_tt_within_radius(int x1, int y1, int x2, int y2, point const &pos, float radius, bool is_square, float mesh_dz=-1.0, tile_t const *const cur_tile=nullptr);
-	unsigned get_gpu_mem() const {return (palm_vbo_mem + vbo_manager[0].get_gpu_mem() + vbo_manager[1].get_gpu_mem());}
+	size_t get_gpu_mem() const {return (palm_vbo_mem + vbo_manager[0].get_gpu_mem() + vbo_manager[1].get_gpu_mem());}
 	bool is_uploaded(bool low_detail) const {return vbo_manager[low_detail].is_uploaded();}
 	void update_zmax(float &tzmax) const;
 	bool update_zvals(int x1, int y1, int x2, int y2);
